Boston Housing Project Machine Learning Practice
This is the project we practice during the Machine learning course.
The Boston Housing dataset is a popular dataset commonly used in regression tasks and predictive modeling. It was collected by Harrison and Rubinfeld in 1978 and consists of various attributes of houses in the Boston area, along with their corresponding median values.

The dataset contains 506 instances and 14 features, including information like an average number of rooms per dwelling, crime rate, accessibility to highways, and more. The target variable, the median value of owner-occupied homes (in thousands of dollars), is often used as the target for regression models.

It has been a staple dataset in the field of machine learning and serves as a benchmark for evaluating regression algorithms and techniques.